chore: Code cleanups
- Replace org.assertj.core.util.Lists.emptyList() by
Collections.emptyList()
- Remove unnecessary casts
- Remove unused imports
- Use switch expressions
- pom.xml
  - remove redundant version for clirr-plugin

+++ engine/src/test/java/org/operaton/bpm/engine/test/api/mgmt/metrics/AbstractMetricsIntervalTest.java
@@ -92,7 +92,7 @@ public abstract class AbstractMetricsIntervalTest {
   protected void reportMetrics() {
     for (String metricName : metricsRegistry.getDbMeters().keySet()) {
       //mark random occurence
-      long occurence = (long) (rand.nextInt((MAX_OCCURENCE - MIN_OCCURENCE) + 1) + MIN_OCCURENCE);
+      long occurence = rand.nextInt((MAX_OCCURENCE - MIN_OCCURENCE) + 1) + MIN_OCCURENCE;
       metricsRegistry.markOccurrence(metricName, occurence);
     }
     //report logged metrics
