+++ engine/src/test/java/org/operaton/bpm/engine/test/bpmn/event/compensate/helper/WaitStateUndoService.java
@@ -34,7 +34,7 @@ public class WaitStateUndoService extends AbstractBpmnActivityBehavior implement
     String variableName = (String) counterName.getValue(execution);
     Object variable = execution.getVariable(variableName);
     if(variable == null) {
-      execution.setVariable(variableName, (Integer) 1);
+      execution.setVariable(variableName, 1);
     }
     else  {
       execution.setVariable(variableName, ((Integer)variable)+1);
